requirements.txt
pillow
uvicorn
fastapi
